requirements.txt
networkx==2.1.0
numpy==1.20.1
scikit-learn==1.0.2
scipy==1.2.1
pytorch==1.10
torchsummary==1.5.1
